feat(ffi): switch iterables and core calculations to immutable versions
