+++ ddnnife_ffi/src/ddnnf_mut.rs
@@ -1,5 +1,4 @@
 use crate::Ddnnf;
-use ddnnife::util;
 use num::BigInt;
 use std::sync::Mutex;
 
@@ -25,60 +24,10 @@ impl DdnnfMut {
         self.0.lock().unwrap().0.execute_query(assumptions)
     }
 
-    /// Computes the cardinality of this d-DNNF for multiple iterables.
-    #[uniffi::method]
-    fn count_iterables(&self, assumptions: &[i32], iterables: &[i32]) -> Vec<BigInt> {
-        self.0
-            .lock()
-            .unwrap()
-            .0
-            .count_iterables(assumptions, iterables)
-    }
-
     /// Computes whether this d-DNNF is satisfiable.
     #[uniffi::method]
     fn is_sat(&self, assumptions: &[i32]) -> bool {
         self.0.lock().unwrap().0.sat(assumptions)
-    }
-
-    /// Computes the core features of this d-DNNF.
-    #[uniffi::method]
-    fn core(&self, assumptions: &[i32]) -> Vec<i32> {
-        self.0.lock().unwrap().0.core_with_assumptions(assumptions)
-    }
-
-    /// Computes the core features of this d-DNNF for multiple variables.
-    #[uniffi::method]
-    fn core_multiple(&self, assumptions: &[i32], variables: &[i32]) -> Vec<i32> {
-        let mut ddnnf = self.0.lock().unwrap();
-
-        let mut core: Vec<i32> = util::zip_assumptions_variables(assumptions, variables)
-            .flat_map(|assumptions| ddnnf.0.core_with_assumptions(&assumptions).into_iter())
-            .collect();
-
-        core.sort();
-        core.dedup();
-        core
-    }
-
-    /// Computes the dead features of this d-DNNF.
-    #[uniffi::method]
-    fn dead(&self, assumptions: &[i32]) -> Vec<i32> {
-        self.0.lock().unwrap().0.dead_with_assumptions(assumptions)
-    }
-
-    /// Computes the dead features of this d-DNNF for multiple variables.
-    #[uniffi::method]
-    fn dead_multiple(&self, assumptions: &[i32], variables: &[i32]) -> Vec<i32> {
-        let mut ddnnf = self.0.lock().unwrap();
-
-        let mut dead: Vec<i32> = util::zip_assumptions_variables(assumptions, variables)
-            .flat_map(|assumptions| ddnnf.0.dead_with_assumptions(&assumptions).into_iter())
-            .collect();
-
-        dead.sort();
-        dead.dedup();
-        dead
     }
 
     /// Generates satisfiable configurations for this d-DNNF.
